moved data and unified_meta properties to the base grouped alert class
