+++ elementary/monitor/alerts/grouped_alerts/grouped_by_table.py
@@ -1,7 +1,6 @@
-from typing import Dict, List, Optional
+from typing import Optional
 
 from elementary.monitor.alerts.grouped_alerts.grouped_alert import GroupedAlert
-from elementary.monitor.alerts.model_alert import ModelAlertModel
 from elementary.monitor.data_monitoring.alerts.integrations.utils.report_link import (
     ReportLinkData,
     get_model_test_runs_link,
@@ -21,26 +20,6 @@ class GroupedByTableAlerts(GroupedAlert):
     @property
     def report_url(self) -> Optional[str]:
         return self.alerts[0].report_url
-
-    @property
-    def unified_meta(self) -> Dict:
-        # If a model level unified meta is defined, we use is.
-        # Else we use one of the tests level unified metas.
-        model_unified_meta = dict()
-        test_unified_meta = dict()
-        for alert in self.alerts:
-            alert_unified_meta = alert.unified_meta
-            if alert_unified_meta:
-                if isinstance(alert, ModelAlertModel):
-                    model_unified_meta = alert_unified_meta
-                    break
-
-                test_unified_meta = alert_unified_meta
-        return model_unified_meta or test_unified_meta
-
-    @property
-    def data(self) -> List[Dict]:
-        return [alert.data for alert in self.alerts]
 
     @property
     def summary(self) -> str:
